+++ concordium-sdk/src/main/java/com/concordium/sdk/Client.java
@@ -16,6 +16,7 @@ import com.concordium.sdk.responses.cryptographicparameters.CryptographicParamet
 import com.concordium.sdk.responses.transactionstatus.TransactionStatus;
 import com.concordium.sdk.transactions.*;
 import com.google.protobuf.ByteString;
+import com.vdurmont.semver4j.Semver;
 import concordium.ConcordiumP2PRpc;
 import concordium.P2PGrpc;
 import io.grpc.ManagedChannel;
@@ -243,6 +244,15 @@ public final class Client {
             throw BlockNotFoundException.from(blockHash);
         }
         return cryptographicParameters;
+    }
+
+    /**
+     * Gets the Semantic Version of the Peer Software / Node
+     * @return {@link Semver} of the Peer / Node
+     */
+    public Semver getVersion() {
+        val value = server().peerVersion(ConcordiumP2PRpc.Empty.newBuilder().build()).getValue();
+        return new Semver(value);
     }
 
     /**